refactor: move getEntryType to util
