+++ packages/astro/src/content/vite-plugin-content-imports.ts
@@ -7,7 +7,7 @@ import { AstroErrorData } from '../core/errors/errors-data.js';
 import { AstroError } from '../core/errors/errors.js';
 import { escapeViteEnvReferences, getFileInfo } from '../vite-plugin-utils/index.js';
 import { contentFileExts, CONTENT_FLAG } from './consts.js';
-import { getEntryType } from './types-generator.js';
+import { getEntryType } from './utils.js';
 import {
 	ContentConfig,
 	getContentPaths,
